+++ examples/saga/main.rs
@@ -4,8 +4,10 @@ use futures::lock::Mutex;
 use sqlx::{Pool, Postgres};
 use uuid::Uuid;
 
-use esrs::postgres::{PgStore, PgStoreBuilder};
-use esrs::{AggregateManager, AggregateState, EventStore};
+use esrs::manager::AggregateManager;
+use esrs::store::postgres::{PgStore, PgStoreBuilder};
+use esrs::store::EventStore;
+use esrs::AggregateState;
 
 use crate::aggregate::{SagaAggregate, SagaCommand, SagaEvent};
 use crate::common::new_pool;
