+++ src/bus/rabbit/error.rs
