+++ src/bus/rabbit/mod.rs
@@ -9,8 +9,9 @@ use serde::Serialize;
 pub use config::RabbitEventBusConfig;
 pub use error::RabbitEventBusError;
 
-use crate::esrs::event_bus::EventBus;
-use crate::{Aggregate, StoreEvent};
+use crate::bus::EventBus;
+use crate::store::StoreEvent;
+use crate::Aggregate;
 
 mod config;
 mod error;
