+++ src/sql/postgres/statements/delete_by_aggregate_id.sql
@@ -1,0 +1,1 @@
+DELETE FROM {} WHERE aggregate_id = $1
