+++ src/sql/postgres/statements/select_by_aggregate_id.sql
@@ -1,0 +1,1 @@
+SELECT * FROM {} WHERE aggregate_id = $1 ORDER BY sequence_number ASC
